Update 4/1: Interim buildable version of Init and Run.
This commit is very tasty. It includes the bulk of the HCO_Init and HCO_Run code, although it has some troubles running.

Warning: This code has a very ugly kludge for a regridding boundary problem that should be looked into with more caution. The kludge is documented in hco_esmf_grid.F90.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>