Update 4/16: Now allow exporting LANDTYPE and XLAI chemistry input data to pbuf.
At request from Thibaud Fritz, HEMCO_CESM now detects CHEMISTRY_INPUTS data from HEMCO_Config.rc and exports to pbuf as appropriate for CESM2-GC.
This is based on magic detection of a container called LANDTYPE00 used in olson landmap. If in the future the landscape changes, the hardcode will need to be adapted.

Additionally, all exports are 3-D right now, even for 2-D data. This needs to be rectified in a future update.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>